Update: Change recent activity folder location
