+++ nexus/event_domain/recent_activity/recent_activities_dto.py
@@ -2,6 +2,7 @@ from dataclasses import dataclass
 from nexus.intelligences.models import Intelligence
 from nexus.projects.models import Project
 from nexus.users.models import User
+from nexus.orgs.models import Org
 
 
 @dataclass
@@ -11,3 +12,12 @@ class CreateRecentActivityDTO:
     created_by: User
     intelligence: Intelligence
     action_details: dict = None
+
+
+@dataclass
+class RecentActivitiesDTO:
+    org: Org
+    user: User
+    entity_name: str
+    action: str
+    entity: str = "NEXUS"
