--- nexus/usecases/event_driven/__init__.py
@@ -1,2 +1,0 @@
-from .recent_activities import intelligence_activity_message
-from .publishers_dto import RecentActivitiesDTO
